+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Version 2.99.x and version 3.x will be available with a [pro license](http://www
 **Next release**
 
 * Show if application has specific fake values ([issue](/../../issues/1831))
+* Allow own package name for *Srv_getPackageInfo* and *Srv_getApplicationInfo*
 * Updated Dutch translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
@@ -198,6 +199,7 @@ XPrivacy 2
 **Next release**
 
 * Show if application has specific fake values ([issue](/../../issues/1831))
+* Allow own package name for *Srv_getPackageInfo* and *Srv_getApplicationInfo*
 * Updated Dutch translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)